Major changes to "View Login History" and "View Page View History". Both pages were grossly inefficient -- loading *all* the data from the database all at once, and then dumping it all to the screen only after it was all down.  Login History was even worse, because then all this data got rendered in a tablesorter.  Load times of five seconds were not uncommon, especially near the end of the semester.
Now we load a small amount of data, and AJAX more when the user
scrolls down to the bottom of the page -- which is the only time it
matters whether that data has been downloaded or not.

ALSO SIGNIFICANT in this update is that, for the first time, the CSS
for a page (Page Views in this case) is in its own file, in the css/
directory.  The default.css file has become totally undreadable and
unnavigable.  I will be making the effort starting today to get as
much individual page CSS out of that file as I can, and into
individual files in the css/ directory, so that searching and editing
can be a lot easier.