OK, this is a *huge* update.  Current users updating from a previous version *must* alter the structure of table student_x_section and add the table student_statuses.
The reason for the change is to accommodate a change in the way Nassau
Community College handles students who don't officially withdraw from
a class for which they registered.  Without getting into it here,
there are three new grades that mean either F or W, and so just having
active or not active wasn't going to cut it for Fall 2012.  So, this
version of the software requires a change to student_x_section, as
noted above.  Instead of a binary condition for active/not active,
there's now a foreign key into a table of statuses, and so you at your
institution can implement as many or as few statuses as you like.  (It
is *strongly* recommended that your statuses include "Grade", "Audit"
and "INC", because I've done some hard-coding to recognize them.)

Anyway, there are a bunch of scripts that were checking for and/or
altering the active/not active status, that had to be updated to
check/alter this new status scheme, and so that's why so many files
are changed in this update.