My nomination for dumbest bug of the year:
graphs/index.php: Sometimes the top graph displayed, sometimes it didn't.  Couldn't tell why.  Happened to have that page open today, had to pull down the left sidebar, and wouldn't you know it, there was the missing graph.  The two divs had the same id.  But since both divs are AJAX'd in, it was a race condition.

Stupid.