A few updates here:
A change to the students' Reference Materials screen -- it automatically refreshes as you add/enable things, so students will never have to hit the reload button if you add things during class.

A change to the admin Reference Materials screen -- an ok drag-and-drop upload tool.

Comments: When students leave a comment on a homework submission, you'd have to remember to go look to see if there are any.  Now, on the admin page, there's a box above the main accordion with details on the most recent comment, and a link to read it in a dialog box, or to go see all the comments on one page.

Quotes: Like at the bottom of the page.  This page lets you see them all and edit them all.  Still need to add functionality to delete one or add one.